requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
langchain
langchain-openai
azure-storage-blob
azure-identity
azure-search-documents
PyPDF2
faiss-cpu
